backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "argon2-cffi>=23.1.0",
    "fastapi[standard]>=0.115.12",
    "pydantic>=2.11.2",
    "python-ulid>=3.0.0",
]
